[FBCommonHeaders] * LockQueue::Enq() now accepts reference or rvalue(&&) as arguments * Added class ProfilerSimple which is indepenent from any other modules or libraries.(header only) * Fixed bug for recursive spin lock
[FBDebug]
* Cleaned logger - removed all cases that had required multiple evaluation of varargs

[FBFileSystem]
* Multi thread considerrations. Fixed bugs Writing-Reading from multiple threads.

[FBMathLib]
* Added a few utility functions: IsPointIn(), IsContained(), RandomDirectionCone(), Vec4::operator[]

[FBSystemLib]
* Added function: SetWindowPos()

[FBAudioPlayer]
* Remove the sudden lack when playing big audio files(musics)
* Fixed music loop bug

[FBRenderer]
* Coning Cameras

[FBSceneObjectFactory]
* Added ISkySphereListener

[FBEngineFacade]
* Added SkyFacade listener(to catch OnInterpolatinoFinished event.)

[FBUI]
* Fixed scrollerbar bugs - when scrollbar resides in the content window.
* Fixed ListBox bugs - in certain cases some row was recovered from hide state.
* Translation of texts in DropDown component
* Fixed crash in ImageBox component when it used in FrameAnimation
* Added new properties: CLOSE_BY_BACKGROUND_CLICK

+++ FBCommonHeaders/Helpers.h
@@ -139,7 +139,7 @@ namespace fb{
 		return numDeleted;
 	}
 
-	inline unsigned GetStringMemSize(const std::string& str) {
+	inline size_t GetStringMemSize(const std::string& str) {
 		return sizeof(size_t) + str.length();
 	}
